Refactored render.ts into folders for easier navigation
