+++ src/lib/gameModes/regular.ts
@@ -1,18 +1,23 @@
 import type { Game } from "../game"
 import { setCanvasSize, getScreenRect, getScreenCenterPosition } from "../canvas_util"
 import { initKeyControllers, spaceObjectKeyController } from "../input"
-import { add, direction, magnitude, rndfVec2d, rndi, round2dec, sub } from "../math"
+import { add, direction, magnitude, rndfVec2d, rndi, round2dec } from "../math"
 import { bounceSpaceObject, handleDeathExplosion } from "../mechanics"
 import { friction, gravity, handleCollisions } from "../physics"
-import { renderMoon, renderExplosionFrame, renderShip, renderSpaceObjectStatusBar, loadingText, renderViewport, renderInfoText } from "../render"
+import { loadingText } from "../render/render2d"
 import { fpsCounter } from "../time"
 import { GameType, getRenderableObjectCount, SpaceShape, type ServerUpdate, type SpaceObject } from "../types"
-import { getSerVer, initMultiplayer, registerServerUpdate } from "../webSocket"
+import { getSerVer, initMultiplayer, registerServerUpdate } from "../websocket/webSocket"
 import { randomAnyColor } from "../color"
 import { test } from "../test"
 import { explosionDuration } from "../constants"
 import { addDataPoint, getLatestValue, GRAPHS, msPretty, newDataStats, renderGraph, type DataStats } from "../stats"
-import { createSpaceObject, newPhotonLaser, reduceShotSize, reduceSoSize } from "../factory"
+import { createSpaceObject, newPhotonLaser } from "../factory"
+import { reduceShotSize, reduceSoSize } from "../websocket/util"
+import { renderMoon } from "../render/renderDebris"
+import { renderShip } from "../render/renderShip"
+import { renderViewport } from "../render/renderUI"
+import { renderExplosionFrame } from "../render/renderFx"
 
 let numberOfServerObjects = 0
 let ops = 0
