+++ src/lib/websocket/webSocket.ts
@@ -1,6 +1,6 @@
-import type { ServerUpdate, SpaceObject } from "./types"
-import { OIDS_WS_PORT } from "../../server/pub_config"
-import { reduceShotSize, reduceSoSize, soFromValueArray, soToValueArray } from "./factory"
+import type { ServerUpdate, SpaceObject } from "../types"
+import { OIDS_WS_PORT } from "../../../server/pub_config"
+import { reduceShotSize, reduceSoSize, soFromValueArray, soToValueArray } from "./util"
 
 let socket: WebSocket
 let serverVersion = "offline"
